moved mouse module into a separate directory and separated new simple mouse events into their own file
